Se agrega cargar cmbAcompañantes Se agrega panel desplegable para turnos en frmMenu(Acompañantes - Consultorios) Se soluciono problema con CUD y Num Afiliado Se cambio un parametro =)

+++ CapaDTO/SistemaDTO/cls_PacienteDTO.cs
@@ -11,7 +11,7 @@ namespace CapaDTO.SistemaDTO
 
         public int dni_titular { get; set; }
 
-        public int num_afiliado { get; set; }
+        public Int64 num_afiliado { get; set; }
 
         public int? id_tipo_dni { get; set; }
